requirements.txt
snowflake-connector-python>=3.0.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
tabulate>=0.9.0
